Execute webhook _after_ leaderboard sync
I predict that the leaderboard sync function throws an error, causing the cron job to repeat itself.

If this does not do the trick (resolve the error), my 2nd guess is that I need to somehow (possibly invasively) store the latest successful webhook execution timestamp in Deno Kv storage to prevent duplicate executions.